PLAT-338: proper AsOf of VStateReport that is sent by VE(P-1), various tests fixes. And SMReport now lives exactly one pulse after VStateReport is sent
